switch case вывод пользователю

+++ src/main/java/com/example/actualcoursebot/model/CurrencyModel.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class CurrencyModel {
     String symbol;
     Double price;
-    Integer timestamp;
+    Long timestamp;
 }
